requirements.txt
# Streamlit frontend dependencies
streamlit>=1.31.0
pandas>=1.5.0
requests>=2.28.0
plotly>=5.0.0
asyncio>=3.4.3

# FastAPI backend dependencies
fastapi>=0.95.0
uvicorn>=0.18.2
prisma>=0.19.1
python-jose>=3.3.0
passlib[bcrypt]>=1.7.4

# PostgreSQL dependencies
asyncpg>=0.27.0

# For background task handling (Optional)
asyncio>=3.4.3

# CORS middleware for FastAPI
fastapi.middleware.cors>=0.2.0

# OpenAI API client (if using GPT models or similar)
openai>=0.27.0
